client/requirements.txt
pyaudio
requests
python-dotenv
pydub
